Fixes #1056: moving and aligning archetypes
