+++ app/extension/archetype/spring-boot/src/main/resources/archetype-resources/src/main/resources/META-INF/spring.factories
