+++ app/extension/archetype/spring-boot/src/main/resources/archetype-resources/src/test/java/__extension-name__ExtensionTest.java
