+++ app/extension/archetype/xml/src/main/resources/archetype-resources/src/main/resources/META-INF/syndesis/syndesis-extension-definition.json
