+++ app/extension/archetype/xml/src/main/resources/archetype-resources/src/main/resources/camel/__extension-name__.xml
