+++ app/extension/archetype/xml/src/main/resources/archetype-resources/src/test/java/__extension-name__ExtensionTest.java
