+++ app/extension/example/log-step/src/main/java/io/syndesis/extension/LogExtension.java
