+++ app/extension/example/log-step/src/main/resources/META-INF/syndesis/syndesis-extension-definition.json
