+++ app/pom.xml
@@ -76,6 +76,11 @@
 
 		<!-- Plugin versions -->
 		<exec-maven-plugin.version>1.6.0</exec-maven-plugin.version>
+
+		<!-- maven archetypes -->
+		<maven-archetype-packaging.version>2.0</maven-archetype-packaging.version>
+		<archetype-packaging.version>2.2</archetype-packaging.version>
+		<maven-archetype-plugin.version>2.2</maven-archetype-plugin.version>
 	</properties>
 
 	<modules>
