--- app/rest/archetypes/pom.xml
@@ -1,38 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-   Copyright (C) 2016 Red Hat, Inc.
-
-   Licensed under the Apache License, Version 2.0 (the "License");
-   you may not use this file except in compliance with the License.
-   You may obtain a copy of the License at
-
-       http://www.apache.org/licenses/LICENSE-2.0
-
-   Unless required by applicable law or agreed to in writing, software
-   distributed under the License is distributed on an "AS IS" BASIS,
-   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-   See the License for the specific language governing permissions and
-   limitations under the License.
--->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <groupId>io.syndesis</groupId>
-    <artifactId>syndesis-rest-parent</artifactId>
-    <version>1.2-SNAPSHOT</version>
-    <relativePath>../pom.xml</relativePath>
-  </parent>
-
-  <groupId>io.syndesis</groupId>
-  <artifactId>archetypes</artifactId>
-  <name>REST :: Archetypes :: Parent</name>
-  <packaging>pom</packaging>
-
-  <modules>
-    <module>syndesis-archetype-flow-extension-boot</module>
-    <module>syndesis-archetype-flow-extension-java</module>
-    <module>syndesis-archetype-flow-extension-xml</module>
-  </modules>
-</project>
