--- app/rest/archetypes/syndesis-archetype-flow-extension-java/src/main/resources-filtered/META-INF/maven/archetype-metadata.xml
@@ -1,73 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-   Copyright (C) 2016 Red Hat, Inc.
-
-   Licensed under the Apache License, Version 2.0 (the "License");
-   you may not use this file except in compliance with the License.
-   You may obtain a copy of the License at
-
-       http://www.apache.org/licenses/LICENSE-2.0
-
-   Unless required by applicable law or agreed to in writing, software
-   distributed under the License is distributed on an "AS IS" BASIS,
-   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-   See the License for the specific language governing permissions and
-   limitations under the License.
--->
-<archetype-descriptor name="basic">
-  <id>syndesis-archetype-flow-extension-java</id>
-
-  <requiredProperties>
-    <requiredProperty key="extension-name"/>
-    <requiredProperty key="extension-description"/>
-    <requiredProperty key="camel-version">
-      <defaultValue>${camel.version}</defaultValue>
-    </requiredProperty>
-    <requiredProperty key="spring-boot-version">
-      <defaultValue>${spring-boot.version}</defaultValue>
-    </requiredProperty>
-    <requiredProperty key="syndesis-version">
-      <defaultValue>${project.version}</defaultValue>
-    </requiredProperty>
-    <requiredProperty key="syndesis-integration-runtime-version">
-      <defaultValue>${syndesis-integration-runtime.version}</defaultValue>
-    </requiredProperty>
-    <requiredProperty key="maven-compiler-plugin-version">
-      <defaultValue>3.7.0</defaultValue>
-    </requiredProperty>
-  </requiredProperties>
-
-  <fileSets>
-    <fileSet filtered="true" packaged="true" encoding="UTF-8">
-      <directory>src/main/java</directory>
-      <includes>
-        <include>**/*.java</include>
-      </includes>
-    </fileSet>
-    <fileSet filtered="true" encoding="UTF-8">
-      <directory>src/main/resources</directory>
-      <includes>
-        <include>**/*</include>
-      </includes>
-    </fileSet>
-    <fileSet filtered="true" packaged="true" encoding="UTF-8">
-      <directory>src/test/java</directory>
-      <includes>
-        <include>**/*.java</include>
-      </includes>
-    </fileSet>
-    <fileSet filtered="true" encoding="UTF-8">
-      <directory>src/test/resources</directory>
-      <includes>
-        <include>**/*</include>
-      </includes>
-    </fileSet>
-    <fileSet encoding="UTF-8">
-      <directory></directory>
-      <includes>
-        <include>README.adoc</include>
-      </includes>
-    </fileSet>
-  </fileSets>
-
-</archetype-descriptor>
