--- app/rest/archetypes/syndesis-archetype-flow-extension-java/src/main/resources/archetype-resources/README.adoc
@@ -1,12 +1,0 @@
-= Syndesis Flow Extension
-
-This project is a template of a Syndesis Flow Extension.
-
-To build this project use
-
-    mvn install
-
-For more help see the Sydnesis documentsion:
-
-    hhttps://syndesis.io/
-
