+++ app/runtime/runtime/pom.xml
@@ -38,8 +38,9 @@
     </dependency>
 
     <dependency>
-      <groupId>io.syndesis.integration-runtime</groupId>
-      <artifactId>runtime-api</artifactId>
+      <groupId>io.syndesis.extension</groupId>
+      <artifactId>extension-api</artifactId>
+      <version>${project.version}</version>
     </dependency>
 
     <dependency>
